done thong ke

+++ database/migrations/2021_02_24_064759_create_products_table.php
@@ -16,8 +16,8 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('price')->nullable();
-            $table->float('sale_price')->nullable();
+            $table->double('price')->nullable();
+            $table->double('sale_price')->nullable();
             $table->bigInteger('inventory_quantity')->nullable();
             $table->text('description')->nullable();
             $table->text('notes')->nullable();
